pyproject.toml
[tool.poetry]
name = "mentor_scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.1.0"
bs4 = "^0.0.1"
requests = "^2.27.1"
pandas = "^1.4.1"
lxml = "^4.7.1"
webdriver-manager = "^3.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
